PartyName appears on all thumbnails on hover
Added an id to every thumbnail image and div containing the
corresponding party name with the same number so that the corresponding
name appears with the different images. Then also added a light white
shadow on the party names when they’re hovered over since the lighter
background disappears when the mouse is only over the name.